cast to int

+++ include/runner.hpp
@@ -90,7 +90,8 @@ public:
                 }
             });
             rep->end_report();
-            return std::min(rep->faults(), static_cast<std::size_t>(std::numeric_limits<std::int32_t>::max()));
+            return static_cast<std::int32_t>(
+              std::min(rep->faults(), static_cast<std::size_t>(std::numeric_limits<std::int32_t>::max())));
         } catch (std::runtime_error const& e) {
             return err_exit(e.what());
         }
